Paint grid
This introduces some new functionalities to the grid to be able to
paint it dynamically using an external color variable from js to css
with the var satement. Now, we need to be able to choose between
painting each pixel individually or click and drag until you click
again. Gotta have a color pallete to choose as well as the rainbow
mode and the erase mode.