Fix event listener logic
Put the gridBlocks declaration inside the event listener. This way
every time you actiavte the click event on an element, it will update
the variable. Nothing worked if not for this fix. Still want to
implement a way tp display clear on the grid division button when it
is not selected.